requirements.txt
absl-py==0.9.0
appdirs==1.4.4
apturl==0.5.2
arrow==0.15.8
asgiref==3.2.10
astunparse==1.6.3
bcrypt==3.1.7
beautifulsoup4==4.9.1
blinker==1.4
Brlapi==0.7.0
cachetools==4.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.0.8
django-bootstrap4==2.2.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gast==0.3.3
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
honcho==1.0.1
httplib2==0.14.0
idna==2.8
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.0
matplotlib==3.3.0
monotonic==1.5
multi==0.0.1
multitimer==0.2
netifaces==0.10.4
numpy==1.19.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.3.0.36
opt-einsum==3.3.0
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pocketsphinx==0.1.15
pool==0.1.2.dev0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-time==0.1.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.6
scipy==1.4.1
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
sqlparse==0.3.1
systemd-python==234
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.27
wadllib==1.3.3
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0

django-filter~=2.3.0